Dissolve emergence_markers.rs
